@@ -3,10 +3,11 @@
     <PropertyGroup>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <LangVersion>default</LangVersion>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
     </PropertyGroup>
     
     <ItemGroup>
+        <ProjectReference Include="..\..\Findx.IP\Findx.IP.csproj" />
         <ProjectReference Include="..\..\framework\Findx.AspNetCore\Findx.AspNetCore.csproj" />
         <ProjectReference Include="..\..\framework\Findx.Castle\Findx.Castle.csproj" />
         <ProjectReference Include="..\..\framework\Findx.DinkToPdf\Findx.DinkToPdf.csproj" />
