@@ -4,11 +4,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Findx.Exceptions;
+using Findx.Common;
 using Findx.Extensions;
-using Findx.Responses;
 using Findx.RulesEngine.Actions;
-using Findx.Utilities;
 using Microsoft.Extensions.Logging;
 using MicrosoftRulesEngine::RulesEngine.Actions;
 using MicrosoftRulesEngine::RulesEngine.Models;
