@@ -3,6 +3,4 @@
 /// <summary>
 ///     度量客户端统计信息接口
 /// </summary>
-public interface IMetricsClientStats
-{
-}
+public interface IMetricsClientStats;
