@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Findx.AspNetCore.Mvc;
-using Findx.Builders;
+using Findx.Core.Builders;
 using Findx.Data;
 using Findx.Extensions;
 using Findx.Jobs;
@@ -14,7 +14,7 @@ using Findx.Machine.Cpu;
 using Findx.Machine.Memory;
 using Findx.Machine.Network;
 using Findx.Machine.System;
-using Findx.RPC;
+using Findx.NewId;
 using Findx.Security;
 using Findx.Utilities;
 using Microsoft.AspNetCore.Http;
